+++ CHANGELOG.md
@@ -19,6 +19,7 @@ If MD5 sum is not listed for a certain release then it means that the container 
 ## [Unreleased]
 - add CHANGELOG.md (this file)
   gwas.py: remove ``loci`` step from default ``--analysis`` (#22)
+- implement ``gwas.py --analysis saige`` option, allowing to run SAIGE analysis
 - update software versions:
   ```
   ec089544b13d3eb39f13728f8584dcde  saige.sif   (SAIGE v0.44.6.5)